Removed *BasePath fields from config (#17541)
[upstream:531d46b529d4068c1ac8d4e46b164a9dc3fd9a74]
