+++ pkg/provider/provider.go
@@ -389,178 +389,6 @@ func ProviderConfigure(ctx context.Context, d *schema.ResourceData, p *schema.Pr
 	}
 	config.BatchingConfig = batchCfg
 
-	// Registered products
-	config.AccessApprovalBasePath = transport_tpg.BaseUrl(registry.GetProduct("accessapproval"), &config)
-	config.AccessContextManagerBasePath = transport_tpg.BaseUrl(registry.GetProduct("accesscontextmanager"), &config)
-	config.ActiveDirectoryBasePath = transport_tpg.BaseUrl(registry.GetProduct("activedirectory"), &config)
-	config.AlloydbBasePath = transport_tpg.BaseUrl(registry.GetProduct("alloydb"), &config)
-	config.ApigeeBasePath = transport_tpg.BaseUrl(registry.GetProduct("apigee"), &config)
-	config.ApihubBasePath = transport_tpg.BaseUrl(registry.GetProduct("apihub"), &config)
-	config.ApikeysBasePath = transport_tpg.BaseUrl(registry.GetProduct("apikeys"), &config)
-	config.AppEngineBasePath = transport_tpg.BaseUrl(registry.GetProduct("appengine"), &config)
-	config.ApphubBasePath = transport_tpg.BaseUrl(registry.GetProduct("apphub"), &config)
-	config.ArtifactRegistryBasePath = transport_tpg.BaseUrl(registry.GetProduct("artifactregistry"), &config)
-	config.AssuredWorkloadsBasePath = transport_tpg.BaseUrl(registry.GetProduct("assuredworkloads"), &config)
-	config.BackupDRBasePath = transport_tpg.BaseUrl(registry.GetProduct("backupdr"), &config)
-	config.BeyondcorpBasePath = transport_tpg.BaseUrl(registry.GetProduct("beyondcorp"), &config)
-	config.BiglakeBasePath = transport_tpg.BaseUrl(registry.GetProduct("biglake"), &config)
-	config.BiglakeIcebergBasePath = transport_tpg.BaseUrl(registry.GetProduct("biglakeiceberg"), &config)
-	config.BigQueryBasePath = transport_tpg.BaseUrl(registry.GetProduct("bigquery"), &config)
-	config.BigqueryAnalyticsHubBasePath = transport_tpg.BaseUrl(registry.GetProduct("bigqueryanalyticshub"), &config)
-	config.BigqueryConnectionBasePath = transport_tpg.BaseUrl(registry.GetProduct("bigqueryconnection"), &config)
-	config.BigqueryDatapolicyBasePath = transport_tpg.BaseUrl(registry.GetProduct("bigquerydatapolicy"), &config)
-	config.BigqueryDatapolicyv2BasePath = transport_tpg.BaseUrl(registry.GetProduct("bigquerydatapolicyv2"), &config)
-	config.BigqueryDataTransferBasePath = transport_tpg.BaseUrl(registry.GetProduct("bigquerydatatransfer"), &config)
-	config.BigqueryReservationBasePath = transport_tpg.BaseUrl(registry.GetProduct("bigqueryreservation"), &config)
-	config.BigtableBasePath = transport_tpg.BaseUrl(registry.GetProduct("bigtable"), &config)
-	config.BillingBudgetsBasePath = transport_tpg.BaseUrl(registry.GetProduct("billingbudgets"), &config)
-	config.BinaryAuthorizationBasePath = transport_tpg.BaseUrl(registry.GetProduct("binaryauthorization"), &config)
-	config.BlockchainNodeEngineBasePath = transport_tpg.BaseUrl(registry.GetProduct("blockchainnodeengine"), &config)
-	config.CertificateManagerBasePath = transport_tpg.BaseUrl(registry.GetProduct("certificatemanager"), &config)
-	config.CESBasePath = transport_tpg.BaseUrl(registry.GetProduct("ces"), &config)
-	config.ChronicleBasePath = transport_tpg.BaseUrl(registry.GetProduct("chronicle"), &config)
-	config.CloudAssetBasePath = transport_tpg.BaseUrl(registry.GetProduct("cloudasset"), &config)
-	config.CloudBillingBasePath = transport_tpg.BaseUrl(registry.GetProduct("cloudbilling"), &config)
-	config.CloudBuildBasePath = transport_tpg.BaseUrl(registry.GetProduct("cloudbuild"), &config)
-	config.Cloudbuildv2BasePath = transport_tpg.BaseUrl(registry.GetProduct("cloudbuildv2"), &config)
-	config.ClouddeployBasePath = transport_tpg.BaseUrl(registry.GetProduct("clouddeploy"), &config)
-	config.ClouddomainsBasePath = transport_tpg.BaseUrl(registry.GetProduct("clouddomains"), &config)
-	config.CloudFunctionsBasePath = transport_tpg.BaseUrl(registry.GetProduct("cloudfunctions"), &config)
-	config.Cloudfunctions2BasePath = transport_tpg.BaseUrl(registry.GetProduct("cloudfunctions2"), &config)
-	config.CloudIdentityBasePath = transport_tpg.BaseUrl(registry.GetProduct("cloudidentity"), &config)
-	config.CloudIdsBasePath = transport_tpg.BaseUrl(registry.GetProduct("cloudids"), &config)
-	config.CloudQuotasBasePath = transport_tpg.BaseUrl(registry.GetProduct("cloudquotas"), &config)
-	config.CloudRunBasePath = transport_tpg.BaseUrl(registry.GetProduct("cloudrun"), &config)
-	config.CloudRunV2BasePath = transport_tpg.BaseUrl(registry.GetProduct("cloudrunv2"), &config)
-	config.CloudSchedulerBasePath = transport_tpg.BaseUrl(registry.GetProduct("cloudscheduler"), &config)
-	config.CloudSecurityComplianceBasePath = transport_tpg.BaseUrl(registry.GetProduct("cloudsecuritycompliance"), &config)
-	config.CloudTasksBasePath = transport_tpg.BaseUrl(registry.GetProduct("cloudtasks"), &config)
-	config.ColabBasePath = transport_tpg.BaseUrl(registry.GetProduct("colab"), &config)
-	config.ComposerBasePath = transport_tpg.BaseUrl(registry.GetProduct("composer"), &config)
-	config.ComputeBasePath = transport_tpg.BaseUrl(registry.GetProduct("compute"), &config)
-	config.ContactCenterInsightsBasePath = transport_tpg.BaseUrl(registry.GetProduct("contactcenterinsights"), &config)
-	config.ContainerBasePath = transport_tpg.BaseUrl(registry.GetProduct("container"), &config)
-	config.ContainerAnalysisBasePath = transport_tpg.BaseUrl(registry.GetProduct("containeranalysis"), &config)
-	config.ContainerAttachedBasePath = transport_tpg.BaseUrl(registry.GetProduct("containerattached"), &config)
-	config.ContainerAwsBasePath = transport_tpg.BaseUrl(registry.GetProduct("containeraws"), &config)
-	config.ContainerAzureBasePath = transport_tpg.BaseUrl(registry.GetProduct("containerazure"), &config)
-	config.DatabaseMigrationServiceBasePath = transport_tpg.BaseUrl(registry.GetProduct("databasemigrationservice"), &config)
-	config.DataCatalogBasePath = transport_tpg.BaseUrl(registry.GetProduct("datacatalog"), &config)
-	config.DataflowBasePath = transport_tpg.BaseUrl(registry.GetProduct("dataflow"), &config)
-	config.DataformBasePath = transport_tpg.BaseUrl(registry.GetProduct("dataform"), &config)
-	config.DataFusionBasePath = transport_tpg.BaseUrl(registry.GetProduct("datafusion"), &config)
-	config.DataLineageBasePath = transport_tpg.BaseUrl(registry.GetProduct("datalineage"), &config)
-	config.DataLossPreventionBasePath = transport_tpg.BaseUrl(registry.GetProduct("datalossprevention"), &config)
-	config.DataPipelineBasePath = transport_tpg.BaseUrl(registry.GetProduct("datapipeline"), &config)
-	config.DataplexBasePath = transport_tpg.BaseUrl(registry.GetProduct("dataplex"), &config)
-	config.DataprocBasePath = transport_tpg.BaseUrl(registry.GetProduct("dataproc"), &config)
-	config.DataprocGdcBasePath = transport_tpg.BaseUrl(registry.GetProduct("dataprocgdc"), &config)
-	config.DataprocMetastoreBasePath = transport_tpg.BaseUrl(registry.GetProduct("dataprocmetastore"), &config)
-	config.DatastreamBasePath = transport_tpg.BaseUrl(registry.GetProduct("datastream"), &config)
-	config.DeploymentManagerBasePath = transport_tpg.BaseUrl(registry.GetProduct("deploymentmanager"), &config)
-	config.DeveloperConnectBasePath = transport_tpg.BaseUrl(registry.GetProduct("developerconnect"), &config)
-	config.DialogflowBasePath = transport_tpg.BaseUrl(registry.GetProduct("dialogflow"), &config)
-	config.DialogflowCXBasePath = transport_tpg.BaseUrl(registry.GetProduct("dialogflowcx"), &config)
-	config.DiscoveryEngineBasePath = transport_tpg.BaseUrl(registry.GetProduct("discoveryengine"), &config)
-	config.DNSBasePath = transport_tpg.BaseUrl(registry.GetProduct("dns"), &config)
-	config.DocumentAIBasePath = transport_tpg.BaseUrl(registry.GetProduct("documentai"), &config)
-	config.DocumentAIWarehouseBasePath = transport_tpg.BaseUrl(registry.GetProduct("documentaiwarehouse"), &config)
-	config.EdgecontainerBasePath = transport_tpg.BaseUrl(registry.GetProduct("edgecontainer"), &config)
-	config.EdgenetworkBasePath = transport_tpg.BaseUrl(registry.GetProduct("edgenetwork"), &config)
-	config.EssentialContactsBasePath = transport_tpg.BaseUrl(registry.GetProduct("essentialcontacts"), &config)
-	config.EventarcBasePath = transport_tpg.BaseUrl(registry.GetProduct("eventarc"), &config)
-	config.FilestoreBasePath = transport_tpg.BaseUrl(registry.GetProduct("filestore"), &config)
-	config.FirebaseAppCheckBasePath = transport_tpg.BaseUrl(registry.GetProduct("firebaseappcheck"), &config)
-	config.FirebaseAppHostingBasePath = transport_tpg.BaseUrl(registry.GetProduct("firebaseapphosting"), &config)
-	config.FirebaseDataConnectBasePath = transport_tpg.BaseUrl(registry.GetProduct("firebasedataconnect"), &config)
-	config.FirebaseRemoteConfigBasePath = transport_tpg.BaseUrl(registry.GetProduct("firebaseremoteconfig"), &config)
-	config.FirebaserulesBasePath = transport_tpg.BaseUrl(registry.GetProduct("firebaserules"), &config)
-	config.FirestoreBasePath = transport_tpg.BaseUrl(registry.GetProduct("firestore"), &config)
-	config.GeminiBasePath = transport_tpg.BaseUrl(registry.GetProduct("gemini"), &config)
-	config.GKEBackupBasePath = transport_tpg.BaseUrl(registry.GetProduct("gkebackup"), &config)
-	config.GKEHubBasePath = transport_tpg.BaseUrl(registry.GetProduct("gkehub"), &config)
-	config.GKEHub2BasePath = transport_tpg.BaseUrl(registry.GetProduct("gkehub2"), &config)
-	config.GkeonpremBasePath = transport_tpg.BaseUrl(registry.GetProduct("gkeonprem"), &config)
-	config.HealthcareBasePath = transport_tpg.BaseUrl(registry.GetProduct("healthcare"), &config)
-	config.HypercomputeclusterBasePath = transport_tpg.BaseUrl(registry.GetProduct("hypercomputecluster"), &config)
-	config.IAM2BasePath = transport_tpg.BaseUrl(registry.GetProduct("iam2"), &config)
-	config.IAM3BasePath = transport_tpg.BaseUrl(registry.GetProduct("iam3"), &config)
-	config.IAMBetaBasePath = transport_tpg.BaseUrl(registry.GetProduct("iambeta"), &config)
-	config.IamCredentialsBasePath = transport_tpg.BaseUrl(registry.GetProduct("iamcredentials"), &config)
-	config.IAMWorkforcePoolBasePath = transport_tpg.BaseUrl(registry.GetProduct("iamworkforcepool"), &config)
-	config.IapBasePath = transport_tpg.BaseUrl(registry.GetProduct("iap"), &config)
-	config.IdentityPlatformBasePath = transport_tpg.BaseUrl(registry.GetProduct("identityplatform"), &config)
-	config.IntegrationConnectorsBasePath = transport_tpg.BaseUrl(registry.GetProduct("integrationconnectors"), &config)
-	config.IntegrationsBasePath = transport_tpg.BaseUrl(registry.GetProduct("integrations"), &config)
-	config.KMSBasePath = transport_tpg.BaseUrl(registry.GetProduct("kms"), &config)
-	config.LoggingBasePath = transport_tpg.BaseUrl(registry.GetProduct("logging"), &config)
-	config.LookerBasePath = transport_tpg.BaseUrl(registry.GetProduct("looker"), &config)
-	config.LustreBasePath = transport_tpg.BaseUrl(registry.GetProduct("lustre"), &config)
-	config.ManagedKafkaBasePath = transport_tpg.BaseUrl(registry.GetProduct("managedkafka"), &config)
-	config.MemcacheBasePath = transport_tpg.BaseUrl(registry.GetProduct("memcache"), &config)
-	config.MemorystoreBasePath = transport_tpg.BaseUrl(registry.GetProduct("memorystore"), &config)
-	config.MigrationCenterBasePath = transport_tpg.BaseUrl(registry.GetProduct("migrationcenter"), &config)
-	config.MLEngineBasePath = transport_tpg.BaseUrl(registry.GetProduct("mlengine"), &config)
-	config.ModelArmorBasePath = transport_tpg.BaseUrl(registry.GetProduct("modelarmor"), &config)
-	config.ModelArmorGlobalBasePath = transport_tpg.BaseUrl(registry.GetProduct("modelarmorglobal"), &config)
-	config.MonitoringBasePath = transport_tpg.BaseUrl(registry.GetProduct("monitoring"), &config)
-	config.NetappBasePath = transport_tpg.BaseUrl(registry.GetProduct("netapp"), &config)
-	config.NetworkConnectivityBasePath = transport_tpg.BaseUrl(registry.GetProduct("networkconnectivity"), &config)
-	config.NetworkConnectivityv1BasePath = transport_tpg.BaseUrl(registry.GetProduct("networkconnectivityv1"), &config)
-	config.NetworkManagementBasePath = transport_tpg.BaseUrl(registry.GetProduct("networkmanagement"), &config)
-	config.NetworkSecurityBasePath = transport_tpg.BaseUrl(registry.GetProduct("networksecurity"), &config)
-	config.NetworkServicesBasePath = transport_tpg.BaseUrl(registry.GetProduct("networkservices"), &config)
-	config.NotebooksBasePath = transport_tpg.BaseUrl(registry.GetProduct("notebooks"), &config)
-	config.ObservabilityBasePath = transport_tpg.BaseUrl(registry.GetProduct("observability"), &config)
-	config.OracleDatabaseBasePath = transport_tpg.BaseUrl(registry.GetProduct("oracledatabase"), &config)
-	config.OrgPolicyBasePath = transport_tpg.BaseUrl(registry.GetProduct("orgpolicy"), &config)
-	config.OSConfigBasePath = transport_tpg.BaseUrl(registry.GetProduct("osconfig"), &config)
-	config.OSConfigV2BasePath = transport_tpg.BaseUrl(registry.GetProduct("osconfigv2"), &config)
-	config.OSLoginBasePath = transport_tpg.BaseUrl(registry.GetProduct("oslogin"), &config)
-	config.ParallelstoreBasePath = transport_tpg.BaseUrl(registry.GetProduct("parallelstore"), &config)
-	config.ParameterManagerBasePath = transport_tpg.BaseUrl(registry.GetProduct("parametermanager"), &config)
-	config.ParameterManagerRegionalBasePath = transport_tpg.BaseUrl(registry.GetProduct("parametermanagerregional"), &config)
-	config.PrivatecaBasePath = transport_tpg.BaseUrl(registry.GetProduct("privateca"), &config)
-	config.PrivilegedAccessManagerBasePath = transport_tpg.BaseUrl(registry.GetProduct("privilegedaccessmanager"), &config)
-	config.PublicCABasePath = transport_tpg.BaseUrl(registry.GetProduct("publicca"), &config)
-	config.PubsubBasePath = transport_tpg.BaseUrl(registry.GetProduct("pubsub"), &config)
-	config.PubsubLiteBasePath = transport_tpg.BaseUrl(registry.GetProduct("pubsublite"), &config)
-	config.RecaptchaEnterpriseBasePath = transport_tpg.BaseUrl(registry.GetProduct("recaptchaenterprise"), &config)
-	config.RedisBasePath = transport_tpg.BaseUrl(registry.GetProduct("redis"), &config)
-	config.ResourceManagerBasePath = transport_tpg.BaseUrl(registry.GetProduct("resourcemanager"), &config)
-	config.ResourceManagerV3BasePath = transport_tpg.BaseUrl(registry.GetProduct("resourcemanagerv3"), &config)
-	config.SecretManagerBasePath = transport_tpg.BaseUrl(registry.GetProduct("secretmanager"), &config)
-	config.SecretManagerRegionalBasePath = transport_tpg.BaseUrl(registry.GetProduct("secretmanagerregional"), &config)
-	config.SecureSourceManagerBasePath = transport_tpg.BaseUrl(registry.GetProduct("securesourcemanager"), &config)
-	config.SecurityCenterBasePath = transport_tpg.BaseUrl(registry.GetProduct("securitycenter"), &config)
-	config.SecurityCenterManagementBasePath = transport_tpg.BaseUrl(registry.GetProduct("securitycentermanagement"), &config)
-	config.SecurityCenterV2BasePath = transport_tpg.BaseUrl(registry.GetProduct("securitycenterv2"), &config)
-	config.SecuritypostureBasePath = transport_tpg.BaseUrl(registry.GetProduct("securityposture"), &config)
-	config.ServiceDirectoryBasePath = transport_tpg.BaseUrl(registry.GetProduct("servicedirectory"), &config)
-	config.ServiceManagementBasePath = transport_tpg.BaseUrl(registry.GetProduct("servicemanagement"), &config)
-	config.ServiceNetworkingBasePath = transport_tpg.BaseUrl(registry.GetProduct("servicenetworking"), &config)
-	config.ServiceUsageBasePath = transport_tpg.BaseUrl(registry.GetProduct("serviceusage"), &config)
-	config.SiteVerificationBasePath = transport_tpg.BaseUrl(registry.GetProduct("siteverification"), &config)
-	config.SourceRepoBasePath = transport_tpg.BaseUrl(registry.GetProduct("sourcerepo"), &config)
-	config.SpannerBasePath = transport_tpg.BaseUrl(registry.GetProduct("spanner"), &config)
-	config.SQLBasePath = transport_tpg.BaseUrl(registry.GetProduct("sql"), &config)
-	config.StorageBasePath = transport_tpg.BaseUrl(registry.GetProduct("storage"), &config)
-	config.StorageBatchOperationsBasePath = transport_tpg.BaseUrl(registry.GetProduct("storagebatchoperations"), &config)
-	config.StorageControlBasePath = transport_tpg.BaseUrl(registry.GetProduct("storagecontrol"), &config)
-	config.StorageInsightsBasePath = transport_tpg.BaseUrl(registry.GetProduct("storageinsights"), &config)
-	config.StorageTransferBasePath = transport_tpg.BaseUrl(registry.GetProduct("storagetransfer"), &config)
-	config.TagsBasePath = transport_tpg.BaseUrl(registry.GetProduct("tags"), &config)
-	config.TagsLocationBasePath = transport_tpg.BaseUrl(registry.GetProduct("tagslocation"), &config)
-	config.TranscoderBasePath = transport_tpg.BaseUrl(registry.GetProduct("transcoder"), &config)
-	config.VectorSearchBasePath = transport_tpg.BaseUrl(registry.GetProduct("vectorsearch"), &config)
-	config.VertexAIBasePath = transport_tpg.BaseUrl(registry.GetProduct("vertexai"), &config)
-	config.VmwareengineBasePath = transport_tpg.BaseUrl(registry.GetProduct("vmwareengine"), &config)
-	config.VPCAccessBasePath = transport_tpg.BaseUrl(registry.GetProduct("vpcaccess"), &config)
-	config.WorkbenchBasePath = transport_tpg.BaseUrl(registry.GetProduct("workbench"), &config)
-	config.WorkflowsBasePath = transport_tpg.BaseUrl(registry.GetProduct("workflows"), &config)
-	config.WorkloadIdentityBasePath = transport_tpg.BaseUrl(registry.GetProduct("workloadidentity"), &config)
-	config.WorkstationsBasePath = transport_tpg.BaseUrl(registry.GetProduct("workstations"), &config)
-
 	stopCtx, ok := schema.StopContext(ctx)
 	if !ok {
 		stopCtx = ctx
