+++ pkg/tfplan2cai/transport/getconfig.go
@@ -43,7 +43,6 @@ func NewConfig(ctx context.Context, project, zone, region string, offline bool, 
 		"GOOGLE_IMPERSONATE_SERVICE_ACCOUNT",
 	})
 
-	transport_tpg.ConfigureBasePaths(cfg)
 	if !offline {
 		if client != nil {
 			cfg.Client = client
