+++ pkg/transport/config.go
@@ -214,527 +214,11 @@ type Config struct {
 
 	CustomEndpoints map[string]string
 
-	AccessApprovalBasePath           string
-	AccessContextManagerBasePath     string
-	ActiveDirectoryBasePath          string
-	AlloydbBasePath                  string
-	ApigeeBasePath                   string
-	ApihubBasePath                   string
-	ApikeysBasePath                  string
-	AppEngineBasePath                string
-	ApphubBasePath                   string
-	ArtifactRegistryBasePath         string
-	AssuredWorkloadsBasePath         string
-	BackupDRBasePath                 string
-	BeyondcorpBasePath               string
-	BiglakeBasePath                  string
-	BiglakeIcebergBasePath           string
-	BigQueryBasePath                 string
-	BigqueryAnalyticsHubBasePath     string
-	BigqueryConnectionBasePath       string
-	BigqueryDatapolicyBasePath       string
-	BigqueryDatapolicyv2BasePath     string
-	BigqueryDataTransferBasePath     string
-	BigqueryReservationBasePath      string
-	BigtableBasePath                 string
-	BillingBudgetsBasePath           string
-	BinaryAuthorizationBasePath      string
-	BlockchainNodeEngineBasePath     string
-	CertificateManagerBasePath       string
-	CESBasePath                      string
-	ChronicleBasePath                string
-	CloudAssetBasePath               string
-	CloudBillingBasePath             string
-	CloudBuildBasePath               string
-	Cloudbuildv2BasePath             string
-	ClouddeployBasePath              string
-	ClouddomainsBasePath             string
-	CloudFunctionsBasePath           string
-	Cloudfunctions2BasePath          string
-	CloudIdentityBasePath            string
-	CloudIdsBasePath                 string
-	CloudQuotasBasePath              string
-	CloudRunBasePath                 string
-	CloudRunV2BasePath               string
-	CloudSchedulerBasePath           string
-	CloudSecurityComplianceBasePath  string
-	CloudTasksBasePath               string
-	ColabBasePath                    string
-	ComposerBasePath                 string
-	ComputeBasePath                  string
-	ContactCenterInsightsBasePath    string
-	ContainerBasePath                string
-	ContainerAnalysisBasePath        string
-	ContainerAttachedBasePath        string
-	ContainerAwsBasePath             string
-	ContainerAzureBasePath           string
-	DatabaseMigrationServiceBasePath string
-	DataCatalogBasePath              string
-	DataflowBasePath                 string
-	DataformBasePath                 string
-	DataFusionBasePath               string
-	DataLineageBasePath              string
-	DataLossPreventionBasePath       string
-	DataPipelineBasePath             string
-	DataplexBasePath                 string
-	DataprocBasePath                 string
-	DataprocGdcBasePath              string
-	DataprocMetastoreBasePath        string
-	DatastreamBasePath               string
-	DeploymentManagerBasePath        string
-	DeveloperConnectBasePath         string
-	DialogflowBasePath               string
-	DialogflowCXBasePath             string
-	DiscoveryEngineBasePath          string
-	DNSBasePath                      string
-	DocumentAIBasePath               string
-	DocumentAIWarehouseBasePath      string
-	EdgecontainerBasePath            string
-	EdgenetworkBasePath              string
-	EssentialContactsBasePath        string
-	EventarcBasePath                 string
-	FilestoreBasePath                string
-	FirebaseAppCheckBasePath         string
-	FirebaseAppHostingBasePath       string
-	FirebaseDataConnectBasePath      string
-	FirebaseRemoteConfigBasePath     string
-	FirebaserulesBasePath            string
-	FirestoreBasePath                string
-	GeminiBasePath                   string
-	GKEBackupBasePath                string
-	GKEHubBasePath                   string
-	GKEHub2BasePath                  string
-	GkeonpremBasePath                string
-	HealthcareBasePath               string
-	HypercomputeclusterBasePath      string
-	IAM2BasePath                     string
-	IAM3BasePath                     string
-	IAMBetaBasePath                  string
-	IamCredentialsBasePath           string
-	IAMWorkforcePoolBasePath         string
-	IapBasePath                      string
-	IdentityPlatformBasePath         string
-	IntegrationConnectorsBasePath    string
-	IntegrationsBasePath             string
-	KMSBasePath                      string
-	LoggingBasePath                  string
-	LookerBasePath                   string
-	LustreBasePath                   string
-	ManagedKafkaBasePath             string
-	MemcacheBasePath                 string
-	MemorystoreBasePath              string
-	MigrationCenterBasePath          string
-	MLEngineBasePath                 string
-	ModelArmorBasePath               string
-	ModelArmorGlobalBasePath         string
-	MonitoringBasePath               string
-	NetappBasePath                   string
-	NetworkConnectivityBasePath      string
-	NetworkConnectivityv1BasePath    string
-	NetworkManagementBasePath        string
-	NetworkSecurityBasePath          string
-	NetworkServicesBasePath          string
-	NotebooksBasePath                string
-	ObservabilityBasePath            string
-	OracleDatabaseBasePath           string
-	OrgPolicyBasePath                string
-	OSConfigBasePath                 string
-	OSConfigV2BasePath               string
-	OSLoginBasePath                  string
-	ParallelstoreBasePath            string
-	ParameterManagerBasePath         string
-	ParameterManagerRegionalBasePath string
-	PrivatecaBasePath                string
-	PrivilegedAccessManagerBasePath  string
-	PublicCABasePath                 string
-	PubsubBasePath                   string
-	PubsubLiteBasePath               string
-	RecaptchaEnterpriseBasePath      string
-	RedisBasePath                    string
-	ResourceManagerBasePath          string
-	ResourceManagerV3BasePath        string
-	SecretManagerBasePath            string
-	SecretManagerRegionalBasePath    string
-	SecureSourceManagerBasePath      string
-	SecurityCenterBasePath           string
-	SecurityCenterManagementBasePath string
-	SecurityCenterV2BasePath         string
-	SecuritypostureBasePath          string
-	ServiceDirectoryBasePath         string
-	ServiceManagementBasePath        string
-	ServiceNetworkingBasePath        string
-	ServiceUsageBasePath             string
-	SiteVerificationBasePath         string
-	SourceRepoBasePath               string
-	SpannerBasePath                  string
-	SQLBasePath                      string
-	StorageBasePath                  string
-	StorageBatchOperationsBasePath   string
-	StorageControlBasePath           string
-	StorageInsightsBasePath          string
-	StorageTransferBasePath          string
-	TagsBasePath                     string
-	TagsLocationBasePath             string
-	TranscoderBasePath               string
-	VectorSearchBasePath             string
-	VertexAIBasePath                 string
-	VmwareengineBasePath             string
-	VPCAccessBasePath                string
-	WorkbenchBasePath                string
-	WorkflowsBasePath                string
-	WorkloadIdentityBasePath         string
-	WorkstationsBasePath             string
-
 	RequestBatcherServiceUsage *RequestBatcher
 	RequestBatcherIam          *RequestBatcher
 
 	PreferGlobalEndpoints   bool
 	PreferRegionalEndpoints bool
-}
-
-const AccessApprovalBasePathKey = "AccessApproval"
-const AccessContextManagerBasePathKey = "AccessContextManager"
-const ActiveDirectoryBasePathKey = "ActiveDirectory"
-const AlloydbBasePathKey = "Alloydb"
-const ApigeeBasePathKey = "Apigee"
-const ApihubBasePathKey = "Apihub"
-const ApikeysBasePathKey = "Apikeys"
-const AppEngineBasePathKey = "AppEngine"
-const ApphubBasePathKey = "Apphub"
-const ArtifactRegistryBasePathKey = "ArtifactRegistry"
-const AssuredWorkloadsBasePathKey = "AssuredWorkloads"
-const BackupDRBasePathKey = "BackupDR"
-const BeyondcorpBasePathKey = "Beyondcorp"
-const BiglakeBasePathKey = "Biglake"
-const BiglakeIcebergBasePathKey = "BiglakeIceberg"
-const BigQueryBasePathKey = "BigQuery"
-const BigqueryAnalyticsHubBasePathKey = "BigqueryAnalyticsHub"
-const BigqueryConnectionBasePathKey = "BigqueryConnection"
-const BigqueryDatapolicyBasePathKey = "BigqueryDatapolicy"
-const BigqueryDatapolicyv2BasePathKey = "BigqueryDatapolicyv2"
-const BigqueryDataTransferBasePathKey = "BigqueryDataTransfer"
-const BigqueryReservationBasePathKey = "BigqueryReservation"
-const BigtableBasePathKey = "Bigtable"
-const BillingBudgetsBasePathKey = "BillingBudgets"
-const BinaryAuthorizationBasePathKey = "BinaryAuthorization"
-const BlockchainNodeEngineBasePathKey = "BlockchainNodeEngine"
-const CertificateManagerBasePathKey = "CertificateManager"
-const CESBasePathKey = "CES"
-const ChronicleBasePathKey = "Chronicle"
-const CloudAssetBasePathKey = "CloudAsset"
-const CloudBillingBasePathKey = "CloudBilling"
-const CloudBuildBasePathKey = "CloudBuild"
-const Cloudbuildv2BasePathKey = "Cloudbuildv2"
-const ClouddeployBasePathKey = "Clouddeploy"
-const ClouddomainsBasePathKey = "Clouddomains"
-const CloudFunctionsBasePathKey = "CloudFunctions"
-const Cloudfunctions2BasePathKey = "Cloudfunctions2"
-const CloudIdentityBasePathKey = "CloudIdentity"
-const CloudIdsBasePathKey = "CloudIds"
-const CloudQuotasBasePathKey = "CloudQuotas"
-const CloudRunBasePathKey = "CloudRun"
-const CloudRunV2BasePathKey = "CloudRunV2"
-const CloudSchedulerBasePathKey = "CloudScheduler"
-const CloudSecurityComplianceBasePathKey = "CloudSecurityCompliance"
-const CloudTasksBasePathKey = "CloudTasks"
-const ColabBasePathKey = "Colab"
-const ComposerBasePathKey = "Composer"
-const ComputeBasePathKey = "Compute"
-const ContactCenterInsightsBasePathKey = "ContactCenterInsights"
-const ContainerBasePathKey = "Container"
-const ContainerAnalysisBasePathKey = "ContainerAnalysis"
-const ContainerAttachedBasePathKey = "ContainerAttached"
-const ContainerAwsBasePathKey = "ContainerAws"
-const ContainerAzureBasePathKey = "ContainerAzure"
-const DatabaseMigrationServiceBasePathKey = "DatabaseMigrationService"
-const DataCatalogBasePathKey = "DataCatalog"
-const DataflowBasePathKey = "Dataflow"
-const DataformBasePathKey = "Dataform"
-const DataFusionBasePathKey = "DataFusion"
-const DataLineageBasePathKey = "DataLineage"
-const DataLossPreventionBasePathKey = "DataLossPrevention"
-const DataPipelineBasePathKey = "DataPipeline"
-const DataplexBasePathKey = "Dataplex"
-const DataprocBasePathKey = "Dataproc"
-const DataprocGdcBasePathKey = "DataprocGdc"
-const DataprocMetastoreBasePathKey = "DataprocMetastore"
-const DatastreamBasePathKey = "Datastream"
-const DeploymentManagerBasePathKey = "DeploymentManager"
-const DeveloperConnectBasePathKey = "DeveloperConnect"
-const DialogflowBasePathKey = "Dialogflow"
-const DialogflowCXBasePathKey = "DialogflowCX"
-const DiscoveryEngineBasePathKey = "DiscoveryEngine"
-const DNSBasePathKey = "DNS"
-const DocumentAIBasePathKey = "DocumentAI"
-const DocumentAIWarehouseBasePathKey = "DocumentAIWarehouse"
-const EdgecontainerBasePathKey = "Edgecontainer"
-const EdgenetworkBasePathKey = "Edgenetwork"
-const EssentialContactsBasePathKey = "EssentialContacts"
-const EventarcBasePathKey = "Eventarc"
-const FilestoreBasePathKey = "Filestore"
-const FirebaseAppCheckBasePathKey = "FirebaseAppCheck"
-const FirebaseAppHostingBasePathKey = "FirebaseAppHosting"
-const FirebaseDataConnectBasePathKey = "FirebaseDataConnect"
-const FirebaseRemoteConfigBasePathKey = "FirebaseRemoteConfig"
-const FirebaserulesBasePathKey = "Firebaserules"
-const FirestoreBasePathKey = "Firestore"
-const GeminiBasePathKey = "Gemini"
-const GKEBackupBasePathKey = "GKEBackup"
-const GKEHubBasePathKey = "GKEHub"
-const GKEHub2BasePathKey = "GKEHub2"
-const GkeonpremBasePathKey = "Gkeonprem"
-const HealthcareBasePathKey = "Healthcare"
-const HypercomputeclusterBasePathKey = "Hypercomputecluster"
-const IAM2BasePathKey = "IAM2"
-const IAM3BasePathKey = "IAM3"
-const IAMBetaBasePathKey = "IAMBeta"
-const IamCredentialsBasePathKey = "IamCredentials"
-const IAMWorkforcePoolBasePathKey = "IAMWorkforcePool"
-const IapBasePathKey = "Iap"
-const IdentityPlatformBasePathKey = "IdentityPlatform"
-const IntegrationConnectorsBasePathKey = "IntegrationConnectors"
-const IntegrationsBasePathKey = "Integrations"
-const KMSBasePathKey = "KMS"
-const LoggingBasePathKey = "Logging"
-const LookerBasePathKey = "Looker"
-const LustreBasePathKey = "Lustre"
-const ManagedKafkaBasePathKey = "ManagedKafka"
-const MemcacheBasePathKey = "Memcache"
-const MemorystoreBasePathKey = "Memorystore"
-const MigrationCenterBasePathKey = "MigrationCenter"
-const MLEngineBasePathKey = "MLEngine"
-const ModelArmorBasePathKey = "ModelArmor"
-const ModelArmorGlobalBasePathKey = "ModelArmorGlobal"
-const MonitoringBasePathKey = "Monitoring"
-const NetappBasePathKey = "Netapp"
-const NetworkConnectivityBasePathKey = "NetworkConnectivity"
-const NetworkConnectivityv1BasePathKey = "NetworkConnectivityv1"
-const NetworkManagementBasePathKey = "NetworkManagement"
-const NetworkSecurityBasePathKey = "NetworkSecurity"
-const NetworkServicesBasePathKey = "NetworkServices"
-const NotebooksBasePathKey = "Notebooks"
-const ObservabilityBasePathKey = "Observability"
-const OracleDatabaseBasePathKey = "OracleDatabase"
-const OrgPolicyBasePathKey = "OrgPolicy"
-const OSConfigBasePathKey = "OSConfig"
-const OSConfigV2BasePathKey = "OSConfigV2"
-const OSLoginBasePathKey = "OSLogin"
-const ParallelstoreBasePathKey = "Parallelstore"
-const ParameterManagerBasePathKey = "ParameterManager"
-const ParameterManagerRegionalBasePathKey = "ParameterManagerRegional"
-const PrivatecaBasePathKey = "Privateca"
-const PrivilegedAccessManagerBasePathKey = "PrivilegedAccessManager"
-const PublicCABasePathKey = "PublicCA"
-const PubsubBasePathKey = "Pubsub"
-const PubsubLiteBasePathKey = "PubsubLite"
-const RecaptchaEnterpriseBasePathKey = "RecaptchaEnterprise"
-const RedisBasePathKey = "Redis"
-const ResourceManagerBasePathKey = "ResourceManager"
-const ResourceManagerV3BasePathKey = "ResourceManagerV3"
-const SecretManagerBasePathKey = "SecretManager"
-const SecretManagerRegionalBasePathKey = "SecretManagerRegional"
-const SecureSourceManagerBasePathKey = "SecureSourceManager"
-const SecurityCenterBasePathKey = "SecurityCenter"
-const SecurityCenterManagementBasePathKey = "SecurityCenterManagement"
-const SecurityCenterV2BasePathKey = "SecurityCenterV2"
-const SecuritypostureBasePathKey = "Securityposture"
-const ServiceDirectoryBasePathKey = "ServiceDirectory"
-const ServiceManagementBasePathKey = "ServiceManagement"
-const ServiceNetworkingBasePathKey = "ServiceNetworking"
-const ServiceUsageBasePathKey = "ServiceUsage"
-const SiteVerificationBasePathKey = "SiteVerification"
-const SourceRepoBasePathKey = "SourceRepo"
-const SpannerBasePathKey = "Spanner"
-const SQLBasePathKey = "SQL"
-const StorageBasePathKey = "Storage"
-const StorageBatchOperationsBasePathKey = "StorageBatchOperations"
-const StorageControlBasePathKey = "StorageControl"
-const StorageInsightsBasePathKey = "StorageInsights"
-const StorageTransferBasePathKey = "StorageTransfer"
-const TagsBasePathKey = "Tags"
-const TagsLocationBasePathKey = "TagsLocation"
-const TranscoderBasePathKey = "Transcoder"
-const VectorSearchBasePathKey = "VectorSearch"
-const VertexAIBasePathKey = "VertexAI"
-const VmwareengineBasePathKey = "Vmwareengine"
-const VPCAccessBasePathKey = "VPCAccess"
-const WorkbenchBasePathKey = "Workbench"
-const WorkflowsBasePathKey = "Workflows"
-const WorkloadIdentityBasePathKey = "WorkloadIdentity"
-const WorkstationsBasePathKey = "Workstations"
-
-// Generated product base paths
-var DefaultBasePaths = map[string]string{
-	AccessApprovalBasePathKey:           "https://accessapproval.googleapis.com/v1/",
-	AccessContextManagerBasePathKey:     "https://accesscontextmanager.googleapis.com/v1/",
-	ActiveDirectoryBasePathKey:          "https://managedidentities.googleapis.com/v1/",
-	AlloydbBasePathKey:                  "https://alloydb.googleapis.com/v1/",
-	ApigeeBasePathKey:                   "https://apigee.googleapis.com/v1/",
-	ApihubBasePathKey:                   "https://apihub.googleapis.com/v1/",
-	ApikeysBasePathKey:                  "https://apikeys.googleapis.com/v2/",
-	AppEngineBasePathKey:                "https://appengine.googleapis.com/v1/",
-	ApphubBasePathKey:                   "https://apphub.googleapis.com/v1/",
-	ArtifactRegistryBasePathKey:         "https://artifactregistry.googleapis.com/v1/",
-	AssuredWorkloadsBasePathKey:         "https://{{location}}-assuredworkloads.googleapis.com/v1/",
-	BackupDRBasePathKey:                 "https://backupdr.googleapis.com/v1/",
-	BeyondcorpBasePathKey:               "https://beyondcorp.googleapis.com/v1/",
-	BiglakeBasePathKey:                  "https://biglake.googleapis.com/v1/",
-	BiglakeIcebergBasePathKey:           "https://biglake.googleapis.com/",
-	BigQueryBasePathKey:                 "https://bigquery.googleapis.com/bigquery/v2/",
-	BigqueryAnalyticsHubBasePathKey:     "https://analyticshub.googleapis.com/v1/",
-	BigqueryConnectionBasePathKey:       "https://bigqueryconnection.googleapis.com/v1/",
-	BigqueryDatapolicyBasePathKey:       "https://bigquerydatapolicy.googleapis.com/v1/",
-	BigqueryDatapolicyv2BasePathKey:     "https://bigquerydatapolicy.googleapis.com/v2/",
-	BigqueryDataTransferBasePathKey:     "https://bigquerydatatransfer.googleapis.com/v1/",
-	BigqueryReservationBasePathKey:      "https://bigqueryreservation.googleapis.com/v1/",
-	BigtableBasePathKey:                 "https://bigtableadmin.googleapis.com/v2/",
-	BillingBudgetsBasePathKey:           "https://billingbudgets.googleapis.com/v1/",
-	BinaryAuthorizationBasePathKey:      "https://binaryauthorization.googleapis.com/v1/",
-	BlockchainNodeEngineBasePathKey:     "https://blockchainnodeengine.googleapis.com/v1/",
-	CertificateManagerBasePathKey:       "https://certificatemanager.googleapis.com/v1/",
-	CESBasePathKey:                      "https://ces.googleapis.com/v1/",
-	ChronicleBasePathKey:                "https://{{location}}-chronicle.googleapis.com/v1/",
-	CloudAssetBasePathKey:               "https://cloudasset.googleapis.com/v1/",
-	CloudBillingBasePathKey:             "https://cloudbilling.googleapis.com/v1/",
-	CloudBuildBasePathKey:               "https://cloudbuild.googleapis.com/v1/",
-	Cloudbuildv2BasePathKey:             "https://cloudbuild.googleapis.com/v2/",
-	ClouddeployBasePathKey:              "https://clouddeploy.googleapis.com/v1/",
-	ClouddomainsBasePathKey:             "https://domains.googleapis.com/v1/",
-	CloudFunctionsBasePathKey:           "https://cloudfunctions.googleapis.com/v1/",
-	Cloudfunctions2BasePathKey:          "https://cloudfunctions.googleapis.com/v2/",
-	CloudIdentityBasePathKey:            "https://cloudidentity.googleapis.com/v1/",
-	CloudIdsBasePathKey:                 "https://ids.googleapis.com/v1/",
-	CloudQuotasBasePathKey:              "https://cloudquotas.googleapis.com/v1/",
-	CloudRunBasePathKey:                 "https://{{location}}-run.googleapis.com/",
-	CloudRunV2BasePathKey:               "https://run.googleapis.com/v2/",
-	CloudSchedulerBasePathKey:           "https://cloudscheduler.googleapis.com/v1/",
-	CloudSecurityComplianceBasePathKey:  "https://cloudsecuritycompliance.googleapis.com/v1/",
-	CloudTasksBasePathKey:               "https://cloudtasks.googleapis.com/v2/",
-	ColabBasePathKey:                    "https://{{location}}-aiplatform.googleapis.com/v1/",
-	ComposerBasePathKey:                 "https://composer.googleapis.com/v1/",
-	ComputeBasePathKey:                  "https://compute.googleapis.com/compute/v1/",
-	ContactCenterInsightsBasePathKey:    "https://contactcenterinsights.googleapis.com/v1/",
-	ContainerBasePathKey:                "https://container.googleapis.com/v1/",
-	ContainerAnalysisBasePathKey:        "https://containeranalysis.googleapis.com/v1/",
-	ContainerAttachedBasePathKey:        "https://{{location}}-gkemulticloud.googleapis.com/v1/",
-	ContainerAwsBasePathKey:             "https://{{location}}-gkemulticloud.googleapis.com/v1/",
-	ContainerAzureBasePathKey:           "https://{{location}}-gkemulticloud.googleapis.com/v1/",
-	DatabaseMigrationServiceBasePathKey: "https://datamigration.googleapis.com/v1/",
-	DataCatalogBasePathKey:              "https://datacatalog.googleapis.com/v1/",
-	DataflowBasePathKey:                 "https://dataflow.googleapis.com/v1b3/",
-	DataformBasePathKey:                 "https://dataform.googleapis.com/v1/",
-	DataFusionBasePathKey:               "https://datafusion.googleapis.com/v1/",
-	DataLineageBasePathKey:              "https://datalineage.googleapis.com/v1/",
-	DataLossPreventionBasePathKey:       "https://dlp.googleapis.com/v2/",
-	DataPipelineBasePathKey:             "https://datapipelines.googleapis.com/v1/",
-	DataplexBasePathKey:                 "https://dataplex.googleapis.com/v1/",
-	DataprocBasePathKey:                 "https://dataproc.googleapis.com/v1/",
-	DataprocGdcBasePathKey:              "https://dataprocgdc.googleapis.com/v1/",
-	DataprocMetastoreBasePathKey:        "https://metastore.googleapis.com/v1/",
-	DatastreamBasePathKey:               "https://datastream.googleapis.com/v1/",
-	DeploymentManagerBasePathKey:        "https://www.googleapis.com/deploymentmanager/v2/",
-	DeveloperConnectBasePathKey:         "https://developerconnect.googleapis.com/v1/",
-	DialogflowBasePathKey:               "https://dialogflow.googleapis.com/v2/",
-	DialogflowCXBasePathKey:             "https://{{location}}-dialogflow.googleapis.com/v3/",
-	DiscoveryEngineBasePathKey:          "https://{{location}}-discoveryengine.googleapis.com/v1/",
-	DNSBasePathKey:                      "https://dns.googleapis.com/dns/v1/",
-	DocumentAIBasePathKey:               "https://{{location}}-documentai.googleapis.com/v1/",
-	DocumentAIWarehouseBasePathKey:      "https://contentwarehouse.googleapis.com/v1/",
-	EdgecontainerBasePathKey:            "https://edgecontainer.googleapis.com/v1/",
-	EdgenetworkBasePathKey:              "https://edgenetwork.googleapis.com/v1/",
-	EssentialContactsBasePathKey:        "https://essentialcontacts.googleapis.com/v1/",
-	EventarcBasePathKey:                 "https://eventarc.googleapis.com/v1/",
-	FilestoreBasePathKey:                "https://file.googleapis.com/v1/",
-	FirebaseAppCheckBasePathKey:         "https://firebaseappcheck.googleapis.com/v1/",
-	FirebaseAppHostingBasePathKey:       "https://firebaseapphosting.googleapis.com/v1/",
-	FirebaseDataConnectBasePathKey:      "https://firebasedataconnect.googleapis.com/v1/",
-	FirebaseRemoteConfigBasePathKey:     "https://firebaseremoteconfig.googleapis.com/v1/",
-	FirebaserulesBasePathKey:            "https://firebaserules.googleapis.com/v1/",
-	FirestoreBasePathKey:                "https://firestore.googleapis.com/v1/",
-	GeminiBasePathKey:                   "https://cloudaicompanion.googleapis.com/v1/",
-	GKEBackupBasePathKey:                "https://gkebackup.googleapis.com/v1/",
-	GKEHubBasePathKey:                   "https://gkehub.googleapis.com/v1/",
-	GKEHub2BasePathKey:                  "https://gkehub.googleapis.com/v1/",
-	GkeonpremBasePathKey:                "https://gkeonprem.googleapis.com/v1/",
-	HealthcareBasePathKey:               "https://healthcare.googleapis.com/v1/",
-	HypercomputeclusterBasePathKey:      "https://hypercomputecluster.googleapis.com/v1/",
-	IAM2BasePathKey:                     "https://iam.googleapis.com/v2/",
-	IAM3BasePathKey:                     "https://iam.googleapis.com/v3/",
-	IAMBetaBasePathKey:                  "https://iam.googleapis.com/v1/",
-	IamCredentialsBasePathKey:           "https://iamcredentials.googleapis.com/v1/",
-	IAMWorkforcePoolBasePathKey:         "https://iam.googleapis.com/v1/",
-	IapBasePathKey:                      "https://iap.googleapis.com/v1/",
-	IdentityPlatformBasePathKey:         "https://identitytoolkit.googleapis.com/v2/",
-	IntegrationConnectorsBasePathKey:    "https://connectors.googleapis.com/v1/",
-	IntegrationsBasePathKey:             "https://integrations.googleapis.com/v1/",
-	KMSBasePathKey:                      "https://cloudkms.googleapis.com/v1/",
-	LoggingBasePathKey:                  "https://logging.googleapis.com/v2/",
-	LookerBasePathKey:                   "https://looker.googleapis.com/v1/",
-	LustreBasePathKey:                   "https://lustre.googleapis.com/v1/",
-	ManagedKafkaBasePathKey:             "https://managedkafka.googleapis.com/v1/",
-	MemcacheBasePathKey:                 "https://memcache.googleapis.com/v1/",
-	MemorystoreBasePathKey:              "https://memorystore.googleapis.com/v1/",
-	MigrationCenterBasePathKey:          "https://migrationcenter.googleapis.com/v1/",
-	MLEngineBasePathKey:                 "https://ml.googleapis.com/v1/",
-	ModelArmorBasePathKey:               "https://modelarmor.{{location}}.rep.googleapis.com/v1/",
-	ModelArmorGlobalBasePathKey:         "https://modelarmor.googleapis.com/v1/",
-	MonitoringBasePathKey:               "https://monitoring.googleapis.com/",
-	NetappBasePathKey:                   "https://netapp.googleapis.com/v1/",
-	NetworkConnectivityBasePathKey:      "https://networkconnectivity.googleapis.com/v1/",
-	NetworkConnectivityv1BasePathKey:    "https://networkconnectivity.googleapis.com/v1/",
-	NetworkManagementBasePathKey:        "https://networkmanagement.googleapis.com/v1/",
-	NetworkSecurityBasePathKey:          "https://networksecurity.googleapis.com/v1/",
-	NetworkServicesBasePathKey:          "https://networkservices.googleapis.com/v1/",
-	NotebooksBasePathKey:                "https://notebooks.googleapis.com/v1/",
-	ObservabilityBasePathKey:            "https://observability.googleapis.com/v1/",
-	OracleDatabaseBasePathKey:           "https://oracledatabase.googleapis.com/v1/",
-	OrgPolicyBasePathKey:                "https://orgpolicy.googleapis.com/v2/",
-	OSConfigBasePathKey:                 "https://osconfig.googleapis.com/v1/",
-	OSConfigV2BasePathKey:               "https://osconfig.googleapis.com/v2/",
-	OSLoginBasePathKey:                  "https://oslogin.googleapis.com/v1/",
-	ParallelstoreBasePathKey:            "https://parallelstore.googleapis.com/v1/",
-	ParameterManagerBasePathKey:         "https://parametermanager.googleapis.com/v1/",
-	ParameterManagerRegionalBasePathKey: "https://parametermanager.{{location}}.rep.googleapis.com/v1/",
-	PrivatecaBasePathKey:                "https://privateca.googleapis.com/v1/",
-	PrivilegedAccessManagerBasePathKey:  "https://privilegedaccessmanager.googleapis.com/v1/",
-	PublicCABasePathKey:                 "https://publicca.googleapis.com/v1/",
-	PubsubBasePathKey:                   "https://pubsub.googleapis.com/v1/",
-	PubsubLiteBasePathKey:               "https://{{region}}-pubsublite.googleapis.com/v1/admin/",
-	RecaptchaEnterpriseBasePathKey:      "https://recaptchaenterprise.googleapis.com/v1/",
-	RedisBasePathKey:                    "https://redis.googleapis.com/v1/",
-	ResourceManagerBasePathKey:          "https://cloudresourcemanager.googleapis.com/v1/",
-	ResourceManagerV3BasePathKey:        "https://cloudresourcemanager.googleapis.com/v3/",
-	SecretManagerBasePathKey:            "https://secretmanager.googleapis.com/v1/",
-	SecretManagerRegionalBasePathKey:    "https://secretmanager.{{location}}.rep.googleapis.com/v1/",
-	SecureSourceManagerBasePathKey:      "https://securesourcemanager.googleapis.com/v1/",
-	SecurityCenterBasePathKey:           "https://securitycenter.googleapis.com/v1/",
-	SecurityCenterManagementBasePathKey: "https://securitycentermanagement.googleapis.com/v1/",
-	SecurityCenterV2BasePathKey:         "https://securitycenter.googleapis.com/v2/",
-	SecuritypostureBasePathKey:          "https://securityposture.googleapis.com/v1/",
-	ServiceDirectoryBasePathKey:         "https://servicedirectory.googleapis.com/v1/",
-	ServiceManagementBasePathKey:        "https://servicemanagement.googleapis.com/v1/",
-	ServiceNetworkingBasePathKey:        "https://servicenetworking.googleapis.com/v1/",
-	ServiceUsageBasePathKey:             "https://serviceusage.googleapis.com/v1/",
-	SiteVerificationBasePathKey:         "https://www.googleapis.com/siteVerification/v1/",
-	SourceRepoBasePathKey:               "https://sourcerepo.googleapis.com/v1/",
-	SpannerBasePathKey:                  "https://spanner.googleapis.com/v1/",
-	SQLBasePathKey:                      "https://sqladmin.googleapis.com/sql/v1beta4/",
-	StorageBasePathKey:                  "https://storage.googleapis.com/storage/v1/",
-	StorageBatchOperationsBasePathKey:   "https://storagebatchoperations.googleapis.com/v1/",
-	StorageControlBasePathKey:           "https://storage.googleapis.com/v2/",
-	StorageInsightsBasePathKey:          "https://storageinsights.googleapis.com/v1/",
-	StorageTransferBasePathKey:          "https://storagetransfer.googleapis.com/v1/",
-	TagsBasePathKey:                     "https://cloudresourcemanager.googleapis.com/v3/",
-	TagsLocationBasePathKey:             "https://{{location}}-cloudresourcemanager.googleapis.com/v3/",
-	TranscoderBasePathKey:               "https://transcoder.googleapis.com/v1/",
-	VectorSearchBasePathKey:             "https://vectorsearch.googleapis.com/v1/",
-	VertexAIBasePathKey:                 "https://{{region}}-aiplatform.googleapis.com/v1/",
-	VmwareengineBasePathKey:             "https://vmwareengine.googleapis.com/v1/",
-	VPCAccessBasePathKey:                "https://vpcaccess.googleapis.com/v1/",
-	WorkbenchBasePathKey:                "https://notebooks.googleapis.com/v2/",
-	WorkflowsBasePathKey:                "https://workflows.googleapis.com/v1/",
-	WorkloadIdentityBasePathKey:         "https://workloadidentity.googleapis.com/v1/",
-	WorkstationsBasePathKey:             "https://workstations.googleapis.com/v1/",
 }
 
 var DefaultClientScopes = []string{
@@ -1248,183 +732,6 @@ func AreADCCredentialsX509() bool {
 func RemoveBasePathVersion(url string) string {
 	re := regexp.MustCompile(`(?P<base>https?://.*)(?P<version>/[^/]+?/$)`)
 	return re.ReplaceAllString(url, "$1/")
-}
-
-// For a consumer of config.go that isn't a full fledged provider and doesn't
-// have its own endpoint mechanism such as sweepers, init {{service}}BasePath
-// values to a default. After using this, you should call config.LoadAndValidate.
-func ConfigureBasePaths(c *Config) {
-	// Generated Products
-	c.AccessApprovalBasePath = DefaultBasePaths[AccessApprovalBasePathKey]
-	c.AccessContextManagerBasePath = DefaultBasePaths[AccessContextManagerBasePathKey]
-	c.ActiveDirectoryBasePath = DefaultBasePaths[ActiveDirectoryBasePathKey]
-	c.AlloydbBasePath = DefaultBasePaths[AlloydbBasePathKey]
-	c.ApigeeBasePath = DefaultBasePaths[ApigeeBasePathKey]
-	c.ApihubBasePath = DefaultBasePaths[ApihubBasePathKey]
-	c.ApikeysBasePath = DefaultBasePaths[ApikeysBasePathKey]
-	c.AppEngineBasePath = DefaultBasePaths[AppEngineBasePathKey]
-	c.ApphubBasePath = DefaultBasePaths[ApphubBasePathKey]
-	c.ArtifactRegistryBasePath = DefaultBasePaths[ArtifactRegistryBasePathKey]
-	c.AssuredWorkloadsBasePath = DefaultBasePaths[AssuredWorkloadsBasePathKey]
-	c.BackupDRBasePath = DefaultBasePaths[BackupDRBasePathKey]
-	c.BeyondcorpBasePath = DefaultBasePaths[BeyondcorpBasePathKey]
-	c.BiglakeBasePath = DefaultBasePaths[BiglakeBasePathKey]
-	c.BiglakeIcebergBasePath = DefaultBasePaths[BiglakeIcebergBasePathKey]
-	c.BigQueryBasePath = DefaultBasePaths[BigQueryBasePathKey]
-	c.BigqueryAnalyticsHubBasePath = DefaultBasePaths[BigqueryAnalyticsHubBasePathKey]
-	c.BigqueryConnectionBasePath = DefaultBasePaths[BigqueryConnectionBasePathKey]
-	c.BigqueryDatapolicyBasePath = DefaultBasePaths[BigqueryDatapolicyBasePathKey]
-	c.BigqueryDatapolicyv2BasePath = DefaultBasePaths[BigqueryDatapolicyv2BasePathKey]
-	c.BigqueryDataTransferBasePath = DefaultBasePaths[BigqueryDataTransferBasePathKey]
-	c.BigqueryReservationBasePath = DefaultBasePaths[BigqueryReservationBasePathKey]
-	c.BigtableBasePath = DefaultBasePaths[BigtableBasePathKey]
-	c.BillingBudgetsBasePath = DefaultBasePaths[BillingBudgetsBasePathKey]
-	c.BinaryAuthorizationBasePath = DefaultBasePaths[BinaryAuthorizationBasePathKey]
-	c.BlockchainNodeEngineBasePath = DefaultBasePaths[BlockchainNodeEngineBasePathKey]
-	c.CertificateManagerBasePath = DefaultBasePaths[CertificateManagerBasePathKey]
-	c.CESBasePath = DefaultBasePaths[CESBasePathKey]
-	c.ChronicleBasePath = DefaultBasePaths[ChronicleBasePathKey]
-	c.CloudAssetBasePath = DefaultBasePaths[CloudAssetBasePathKey]
-	c.CloudBillingBasePath = DefaultBasePaths[CloudBillingBasePathKey]
-	c.CloudBuildBasePath = DefaultBasePaths[CloudBuildBasePathKey]
-	c.Cloudbuildv2BasePath = DefaultBasePaths[Cloudbuildv2BasePathKey]
-	c.ClouddeployBasePath = DefaultBasePaths[ClouddeployBasePathKey]
-	c.ClouddomainsBasePath = DefaultBasePaths[ClouddomainsBasePathKey]
-	c.CloudFunctionsBasePath = DefaultBasePaths[CloudFunctionsBasePathKey]
-	c.Cloudfunctions2BasePath = DefaultBasePaths[Cloudfunctions2BasePathKey]
-	c.CloudIdentityBasePath = DefaultBasePaths[CloudIdentityBasePathKey]
-	c.CloudIdsBasePath = DefaultBasePaths[CloudIdsBasePathKey]
-	c.CloudQuotasBasePath = DefaultBasePaths[CloudQuotasBasePathKey]
-	c.CloudRunBasePath = DefaultBasePaths[CloudRunBasePathKey]
-	c.CloudRunV2BasePath = DefaultBasePaths[CloudRunV2BasePathKey]
-	c.CloudSchedulerBasePath = DefaultBasePaths[CloudSchedulerBasePathKey]
-	c.CloudSecurityComplianceBasePath = DefaultBasePaths[CloudSecurityComplianceBasePathKey]
-	c.CloudTasksBasePath = DefaultBasePaths[CloudTasksBasePathKey]
-	c.ColabBasePath = DefaultBasePaths[ColabBasePathKey]
-	c.ComposerBasePath = DefaultBasePaths[ComposerBasePathKey]
-	c.ComputeBasePath = DefaultBasePaths[ComputeBasePathKey]
-	c.ContactCenterInsightsBasePath = DefaultBasePaths[ContactCenterInsightsBasePathKey]
-	c.ContainerBasePath = DefaultBasePaths[ContainerBasePathKey]
-	c.ContainerAnalysisBasePath = DefaultBasePaths[ContainerAnalysisBasePathKey]
-	c.ContainerAttachedBasePath = DefaultBasePaths[ContainerAttachedBasePathKey]
-	c.ContainerAwsBasePath = DefaultBasePaths[ContainerAwsBasePathKey]
-	c.ContainerAzureBasePath = DefaultBasePaths[ContainerAzureBasePathKey]
-	c.DatabaseMigrationServiceBasePath = DefaultBasePaths[DatabaseMigrationServiceBasePathKey]
-	c.DataCatalogBasePath = DefaultBasePaths[DataCatalogBasePathKey]
-	c.DataflowBasePath = DefaultBasePaths[DataflowBasePathKey]
-	c.DataformBasePath = DefaultBasePaths[DataformBasePathKey]
-	c.DataFusionBasePath = DefaultBasePaths[DataFusionBasePathKey]
-	c.DataLineageBasePath = DefaultBasePaths[DataLineageBasePathKey]
-	c.DataLossPreventionBasePath = DefaultBasePaths[DataLossPreventionBasePathKey]
-	c.DataPipelineBasePath = DefaultBasePaths[DataPipelineBasePathKey]
-	c.DataplexBasePath = DefaultBasePaths[DataplexBasePathKey]
-	c.DataprocBasePath = DefaultBasePaths[DataprocBasePathKey]
-	c.DataprocGdcBasePath = DefaultBasePaths[DataprocGdcBasePathKey]
-	c.DataprocMetastoreBasePath = DefaultBasePaths[DataprocMetastoreBasePathKey]
-	c.DatastreamBasePath = DefaultBasePaths[DatastreamBasePathKey]
-	c.DeploymentManagerBasePath = DefaultBasePaths[DeploymentManagerBasePathKey]
-	c.DeveloperConnectBasePath = DefaultBasePaths[DeveloperConnectBasePathKey]
-	c.DialogflowBasePath = DefaultBasePaths[DialogflowBasePathKey]
-	c.DialogflowCXBasePath = DefaultBasePaths[DialogflowCXBasePathKey]
-	c.DiscoveryEngineBasePath = DefaultBasePaths[DiscoveryEngineBasePathKey]
-	c.DNSBasePath = DefaultBasePaths[DNSBasePathKey]
-	c.DocumentAIBasePath = DefaultBasePaths[DocumentAIBasePathKey]
-	c.DocumentAIWarehouseBasePath = DefaultBasePaths[DocumentAIWarehouseBasePathKey]
-	c.EdgecontainerBasePath = DefaultBasePaths[EdgecontainerBasePathKey]
-	c.EdgenetworkBasePath = DefaultBasePaths[EdgenetworkBasePathKey]
-	c.EssentialContactsBasePath = DefaultBasePaths[EssentialContactsBasePathKey]
-	c.EventarcBasePath = DefaultBasePaths[EventarcBasePathKey]
-	c.FilestoreBasePath = DefaultBasePaths[FilestoreBasePathKey]
-	c.FirebaseAppCheckBasePath = DefaultBasePaths[FirebaseAppCheckBasePathKey]
-	c.FirebaseAppHostingBasePath = DefaultBasePaths[FirebaseAppHostingBasePathKey]
-	c.FirebaseDataConnectBasePath = DefaultBasePaths[FirebaseDataConnectBasePathKey]
-	c.FirebaseRemoteConfigBasePath = DefaultBasePaths[FirebaseRemoteConfigBasePathKey]
-	c.FirebaserulesBasePath = DefaultBasePaths[FirebaserulesBasePathKey]
-	c.FirestoreBasePath = DefaultBasePaths[FirestoreBasePathKey]
-	c.GeminiBasePath = DefaultBasePaths[GeminiBasePathKey]
-	c.GKEBackupBasePath = DefaultBasePaths[GKEBackupBasePathKey]
-	c.GKEHubBasePath = DefaultBasePaths[GKEHubBasePathKey]
-	c.GKEHub2BasePath = DefaultBasePaths[GKEHub2BasePathKey]
-	c.GkeonpremBasePath = DefaultBasePaths[GkeonpremBasePathKey]
-	c.HealthcareBasePath = DefaultBasePaths[HealthcareBasePathKey]
-	c.HypercomputeclusterBasePath = DefaultBasePaths[HypercomputeclusterBasePathKey]
-	c.IAM2BasePath = DefaultBasePaths[IAM2BasePathKey]
-	c.IAM3BasePath = DefaultBasePaths[IAM3BasePathKey]
-	c.IAMBetaBasePath = DefaultBasePaths[IAMBetaBasePathKey]
-	c.IamCredentialsBasePath = DefaultBasePaths[IamCredentialsBasePathKey]
-	c.IAMWorkforcePoolBasePath = DefaultBasePaths[IAMWorkforcePoolBasePathKey]
-	c.IapBasePath = DefaultBasePaths[IapBasePathKey]
-	c.IdentityPlatformBasePath = DefaultBasePaths[IdentityPlatformBasePathKey]
-	c.IntegrationConnectorsBasePath = DefaultBasePaths[IntegrationConnectorsBasePathKey]
-	c.IntegrationsBasePath = DefaultBasePaths[IntegrationsBasePathKey]
-	c.KMSBasePath = DefaultBasePaths[KMSBasePathKey]
-	c.LoggingBasePath = DefaultBasePaths[LoggingBasePathKey]
-	c.LookerBasePath = DefaultBasePaths[LookerBasePathKey]
-	c.LustreBasePath = DefaultBasePaths[LustreBasePathKey]
-	c.ManagedKafkaBasePath = DefaultBasePaths[ManagedKafkaBasePathKey]
-	c.MemcacheBasePath = DefaultBasePaths[MemcacheBasePathKey]
-	c.MemorystoreBasePath = DefaultBasePaths[MemorystoreBasePathKey]
-	c.MigrationCenterBasePath = DefaultBasePaths[MigrationCenterBasePathKey]
-	c.MLEngineBasePath = DefaultBasePaths[MLEngineBasePathKey]
-	c.ModelArmorBasePath = DefaultBasePaths[ModelArmorBasePathKey]
-	c.ModelArmorGlobalBasePath = DefaultBasePaths[ModelArmorGlobalBasePathKey]
-	c.MonitoringBasePath = DefaultBasePaths[MonitoringBasePathKey]
-	c.NetappBasePath = DefaultBasePaths[NetappBasePathKey]
-	c.NetworkConnectivityBasePath = DefaultBasePaths[NetworkConnectivityBasePathKey]
-	c.NetworkConnectivityv1BasePath = DefaultBasePaths[NetworkConnectivityv1BasePathKey]
-	c.NetworkManagementBasePath = DefaultBasePaths[NetworkManagementBasePathKey]
-	c.NetworkSecurityBasePath = DefaultBasePaths[NetworkSecurityBasePathKey]
-	c.NetworkServicesBasePath = DefaultBasePaths[NetworkServicesBasePathKey]
-	c.NotebooksBasePath = DefaultBasePaths[NotebooksBasePathKey]
-	c.ObservabilityBasePath = DefaultBasePaths[ObservabilityBasePathKey]
-	c.OracleDatabaseBasePath = DefaultBasePaths[OracleDatabaseBasePathKey]
-	c.OrgPolicyBasePath = DefaultBasePaths[OrgPolicyBasePathKey]
-	c.OSConfigBasePath = DefaultBasePaths[OSConfigBasePathKey]
-	c.OSConfigV2BasePath = DefaultBasePaths[OSConfigV2BasePathKey]
-	c.OSLoginBasePath = DefaultBasePaths[OSLoginBasePathKey]
-	c.ParallelstoreBasePath = DefaultBasePaths[ParallelstoreBasePathKey]
-	c.ParameterManagerBasePath = DefaultBasePaths[ParameterManagerBasePathKey]
-	c.ParameterManagerRegionalBasePath = DefaultBasePaths[ParameterManagerRegionalBasePathKey]
-	c.PrivatecaBasePath = DefaultBasePaths[PrivatecaBasePathKey]
-	c.PrivilegedAccessManagerBasePath = DefaultBasePaths[PrivilegedAccessManagerBasePathKey]
-	c.PublicCABasePath = DefaultBasePaths[PublicCABasePathKey]
-	c.PubsubBasePath = DefaultBasePaths[PubsubBasePathKey]
-	c.PubsubLiteBasePath = DefaultBasePaths[PubsubLiteBasePathKey]
-	c.RecaptchaEnterpriseBasePath = DefaultBasePaths[RecaptchaEnterpriseBasePathKey]
-	c.RedisBasePath = DefaultBasePaths[RedisBasePathKey]
-	c.ResourceManagerBasePath = DefaultBasePaths[ResourceManagerBasePathKey]
-	c.ResourceManagerV3BasePath = DefaultBasePaths[ResourceManagerV3BasePathKey]
-	c.SecretManagerBasePath = DefaultBasePaths[SecretManagerBasePathKey]
-	c.SecretManagerRegionalBasePath = DefaultBasePaths[SecretManagerRegionalBasePathKey]
-	c.SecureSourceManagerBasePath = DefaultBasePaths[SecureSourceManagerBasePathKey]
-	c.SecurityCenterBasePath = DefaultBasePaths[SecurityCenterBasePathKey]
-	c.SecurityCenterManagementBasePath = DefaultBasePaths[SecurityCenterManagementBasePathKey]
-	c.SecurityCenterV2BasePath = DefaultBasePaths[SecurityCenterV2BasePathKey]
-	c.SecuritypostureBasePath = DefaultBasePaths[SecuritypostureBasePathKey]
-	c.ServiceDirectoryBasePath = DefaultBasePaths[ServiceDirectoryBasePathKey]
-	c.ServiceManagementBasePath = DefaultBasePaths[ServiceManagementBasePathKey]
-	c.ServiceNetworkingBasePath = DefaultBasePaths[ServiceNetworkingBasePathKey]
-	c.ServiceUsageBasePath = DefaultBasePaths[ServiceUsageBasePathKey]
-	c.SiteVerificationBasePath = DefaultBasePaths[SiteVerificationBasePathKey]
-	c.SourceRepoBasePath = DefaultBasePaths[SourceRepoBasePathKey]
-	c.SpannerBasePath = DefaultBasePaths[SpannerBasePathKey]
-	c.SQLBasePath = DefaultBasePaths[SQLBasePathKey]
-	c.StorageBasePath = DefaultBasePaths[StorageBasePathKey]
-	c.StorageBatchOperationsBasePath = DefaultBasePaths[StorageBatchOperationsBasePathKey]
-	c.StorageControlBasePath = DefaultBasePaths[StorageControlBasePathKey]
-	c.StorageInsightsBasePath = DefaultBasePaths[StorageInsightsBasePathKey]
-	c.StorageTransferBasePath = DefaultBasePaths[StorageTransferBasePathKey]
-	c.TagsBasePath = DefaultBasePaths[TagsBasePathKey]
-	c.TagsLocationBasePath = DefaultBasePaths[TagsLocationBasePathKey]
-	c.TranscoderBasePath = DefaultBasePaths[TranscoderBasePathKey]
-	c.VectorSearchBasePath = DefaultBasePaths[VectorSearchBasePathKey]
-	c.VertexAIBasePath = DefaultBasePaths[VertexAIBasePathKey]
-	c.VmwareengineBasePath = DefaultBasePaths[VmwareengineBasePathKey]
-	c.VPCAccessBasePath = DefaultBasePaths[VPCAccessBasePathKey]
-	c.WorkbenchBasePath = DefaultBasePaths[WorkbenchBasePathKey]
-	c.WorkflowsBasePath = DefaultBasePaths[WorkflowsBasePathKey]
-	c.WorkloadIdentityBasePath = DefaultBasePaths[WorkloadIdentityBasePathKey]
-	c.WorkstationsBasePath = DefaultBasePaths[WorkstationsBasePathKey]
 }
 
 func GetCurrentUserEmail(config *Config, userAgent string) (string, error) {
